Improvement: Transfer business logic to service for Meal Plan (#28)
